pyproject.toml
[tool.poetry]
# [DO NOT MODIFY] info start
authors = [ "whitespace_negative <whinyaan@protonmail.com>",]
maintainers = [ "whitespace_negative <whinyaan@protonmail.com>",]
name = "wh-snips"
license = "MIT"
version = "0.0.0.0-alpha.0"
description = "Bloated python utility library"
homepage = "https://whsn.whinyaan.xyz"
repository = "https://github.com/whinee/whinesnips"
documentation = "https://whsn.whinyaan.xyz"
# info end

packages = [
# [DO NOT MODIFY] package_include_src start
    { include = "whinesnips" }
# package_include_src end
]

classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities"
]
keywords = ["your", "keywords", "here"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<4"

rich = "^13.3.3"
arrow = "^1.2.3"
markdown2 = "^2.4.8"
msgpack = "^1.0.5"
pyyaml = "^6.0"
inquirer = "^3.1.3"
toml = "^0.10.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
inquirer = "^3.1.3"
isort = "^5.12.0"
mako = "^1.2.4"
mdformat = "^0.7.16"
mdformat-frontmatter = "^2.0.1"
mdformat-footnote = "^0.1.1"
mdformat-gfm = "^0.3.5"
mdformat-shfmt = "^0.1.0"
mkdocs-redirects = "^1.2.0"
mypy = "^1.1.1"
no-implicit-optional = "^1.3"
panflute = "^2.3.0"
pdoc3 = "^0.10.0"
pypandoc = "^1.11"
pypandoc_binary = "^1.11"
python-frontmatter = "^1.0.0"
ruff = "^0.0.259"
pytest = "^7.3.2"
igittigitt = "^2.1.4"
pillow = "^10.0.0"
cairosvg = "^2.7.1"
mkdocs-material = {version = "^9.2.0b3", allow-prereleases = true}
twine = "^4.0.2"

[tool.mypy]
disable_error_code = ["import", "attr-defined", "no-untyped-call"]

[tool.ruff]
select = [
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "D",
    # "DTZ",
    "E",
    "F",
    "I",
    "INP",
    "N",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "UP"
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "ANN",
    "B",
    "C4",
    "COM",
    "D",
    "E",
    "F",
    "I",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "UP"
]
unfixable = []

ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN201",
    "ANN202",
    "ANN401",
    "C901",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D202",
    "D203",
    "D212",
    "D401",
    "D406",
    "D407",
    "D417",
    "E501",
    "I002",
    "N812",
    "N818",
    "RET503",
    "S101",
    "UP007",
    "UP032",
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "pyenv",
    "venv",
]

line-length = 88

target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 5

[tool.ruff.pycodestyle]
ignore-overlong-task-comments = false